+++ src/components/index.ts
@@ -5,3 +5,4 @@ export { default as TranslationBox } from "./TranslationBox";
 export { default as SettingsDialog } from "./SettingsDialog";
 export { default as Footer } from "./Footer";
 export { default as Disclaimer } from "./Disclaimer";
+export { default as LlmDebugDialog } from "./LlmDebugDialog";